requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blockbuster==1.5.24
bs4==0.0.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cryptography==44.0.3
distro==1.9.0
fastapi==0.115.13
forbiddenfruit==0.1.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-api==0.2.61
langgraph-checkpoint==2.1.0
langgraph-cli==0.3.3
langgraph-prebuilt==0.2.2
langgraph-runtime-inmem==0.3.0
langgraph-sdk==0.1.70
langsmith==0.4.1
openai==1.90.0
orjson==3.10.18
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==24.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.1.3
starlette==0.46.2
structlog==25.4.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
truststore==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.1.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xxhash==3.5.0
zstandard==0.23.0
